==> 07_C1_merging_keyword_tweets_social_panel.R <==
# 00_analysing_filtered_tweets.R
# script to analyse filtered tweets from the entire panel during wave 01 [Sep 06 - Oct 10]
# for each of the available categories
# NL, 10/10/18
# NL, 13/10/18 
# NL, 30/10/18    updated for wave 02
# NL, 05/11/18    updated for wave 03
# NL, 05/11/18    updated for wave 04
# NL, 05/11/18    updated for wave 05
# NL, 06/11/18    updated for wave 06
# NL, 15/01/18    big update changing functionality to drop the wave-concept and analyse 
#                 all data in one go for the whole timeframe of interest, 
#                 aka 00_analysing_filtered_tweets.R

library(dplyr)

DATA_PATH <- "~/Dropbox/PhD/midterms_paper/tweets/panel/"

# setwd("~/Dropbox/PhD/midterms_paper/tweets/panel/tweets/")


# TO DO:
# - read in all the data that was output from the previous "analysing_filtered_tweets_" scripts
# - concatenate the existing csvs
# - output the new csvs
# - make everything in this new script replicable

# DATA IN
# these were generated by running analysing_filtered_tweets script for each of the waves.
# hence, all waves' file structures must be iterated through to recursively read in
# these files:

# 1 - to train classifiers with. voting ground truth data
# imvoting.csv
# or, more likely: imvoting_coded.csv

# 2 - all tweets that fit inclusion criteria for the relevant topics
# trump_mention_tweets.csv
# kavanaugh_mention_tweets.csv
# republicans_mention_tweets.csv
# democrats_mention_tweets.csv
# midterms_mention_tweets.csv

# 3 - tweets which feature relevant hashtags, and the number of times they occur in the data
# dem_hashtag_tweets.csv
# rep_hashtag_tweets.csv
# all_hashtags_frequencies.csv

# 4 - no idea what this will be, but let's give it a go
# panel_members_relevant_mention_counts_waveXX.csv

imvoting_coded_df <- NULL
trump_mentions_df <- NULL
kavanaugh_mentions_df <- NULL
midterms_mentions_df <- NULL
democrats_mentions_df <- NULL
republican_mentions_df <- NULL
dem_hashtag_tweets_df <- NULL
rep_hashtag_tweets_df <- NULL
all_hashtags_frequencies_df <- NULL

for(i in 1:6){
  
  # voting
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/imvoting_coded.csv"), 
                      colClasses = "character",
                      col.names = c("X", "user_id", "status_id", "created_at", "vote_label", "text")
                      ) %>% distinct(
    status_id, .keep_all = T
  )
  
  imvoting_coded_df <- rbind(imvoting_coded_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # trump
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/trump_mention_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text", "source", "favorite_count"
  )
  
  trump_mentions_df <- rbind(trump_mentions_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # kavanaugh
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/kavanaugh_mention_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text", "source", "favorite_count"
  )
  
  kavanaugh_mentions_df <- rbind(kavanaugh_mentions_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # midterms
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/midterms_mention_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text"
  )
  
  midterms_mentions_df <- rbind(midterms_mentions_df, pos1_df)  %>% distinct(
    status_id, .keep_all = T
  )
  
  # democrats
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/democrats_mention_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text"
  )
  
  democrats_mentions_df <- rbind(democrats_mentions_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # republicans
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/republicans_mention_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text"
  )
  
  republican_mentions_df <- rbind(republican_mentions_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # dem_hashtag_tweets
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/dem_hashtag_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text", "source", "favorite_count"
  )
  
  dem_hashtag_tweets_df <- rbind(dem_hashtag_tweets_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # rep_hashtag_tweets
  pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/rep_hashtag_tweets.csv"), colClasses = "character") %>% distinct(
    status_id, .keep_all = T
  ) %>% select(
    "X", "user_id", "status_id", "created_at", "text", "source", "favorite_count"
  )
  
  rep_hashtag_tweets_df <- rbind(rep_hashtag_tweets_df, pos1_df) %>% distinct(
    status_id, .keep_all = T
  )
  
  # all hashtag frequencies
  # pos1_df <- read.csv(paste0(DATA_PATH, "wave_0", i, "_data/all_hashtags_frequencies.csv"), stringsAsFactors = F) 
  # 
  # if(i==1){
  #   
  #   all_hashtags_frequencies_df <- rbind(all_hashtags_frequencies_df, pos1_df)
  #   
  # } else {
  #   
  #   # all_hashtags_frequencies_df <- bind_rows(all_hashtags_frequencies_df %>% add_rownames(),
  #   #                                          pos1_df %>% add_rownames()) %>% group_by(
  #   #                                            rowname
  #   #                                          ) %>% summarise_if(
  #   #                                            is.numeric, sum,
  #   #                                          )
  #   
  # }

}

# write out

if(!dir.exists(paste0(DATA_PATH, "wave_everything_data"))){
  dir.create(paste0(DATA_PATH, "wave_everything_data"))
}

write.csv(imvoting_coded_df, paste0(DATA_PATH, "wave_everything_data/imvoting_coded.csv"))
write.csv(trump_mentions_df, paste0(DATA_PATH, "wave_everything_data/trump_mention_tweets.csv"))
write.csv(kavanaugh_mentions_df, paste0(DATA_PATH, "wave_everything_data/kavanaugh_mention_tweets.csv"))
write.csv(midterms_mentions_df, paste0(DATA_PATH, "wave_everything_data/midterms_mention_tweets.csv"))
write.csv(democrats_mentions_df, paste0(DATA_PATH, "wave_everything_data/democrats_mention_tweets.csv"))
write.csv(republican_mentions_df, paste0(DATA_PATH, "wave_everything_data/republicans_mention_tweets.csv"))
write.csv(dem_hashtag_tweets_df, paste0(DATA_PATH, "wave_everything_data/dem_hashtag_tweets.csv"))
write.csv(rep_hashtag_tweets_df, paste0(DATA_PATH, "wave_everything_data/rep_hashtag_tweets.csv"))
